refactor(scripts): Rename scripts
