+++ republica-relatorio/src/script/postgresqlQuerys.ts
